requirements.txt
asgiref==3.3.4
Django==3.2
gunicorn==20.1.0
djangorestframework==3.12.4
pytz==2021.1
python-dotenv==0.12.0
redis==3.5.3
sqlparse==0.4.1
typing-extensions==3.7.4.3
